requirements.txt
Google-Images-Search
virtualenv
python-dotenv
windows-curses